+++ lib/http_wrapper.rb
@@ -1,5 +1,6 @@
 require 'http_wrapper/version'
 require 'http_wrapper/constants'
 require 'http_wrapper/errors'
+require 'http_wrapper/utils'
 require 'http_wrapper/request'
 require 'http_wrapper/http_wrapper'