+++ lib/http_wrapper/errors.rb
@@ -1,5 +1,5 @@
 class HTTPWrapper
   HTTPWrapperError      = Class.new StandardError
   TooManyRedirectsError = Class.new HTTPWrapperError
-  UnknownParameterError = Class.new HTTPWrapperError
+  UnknownKeyError       = Class.new HTTPWrapperError
 end